requirements.txt
langchain==0.0.154
pypdf2==3.0.1
python-dotenv~=1.0.1
streamlit~=1.34.0
faiss-cpu==1.7.4
groq==0.4.0
tabula-py
tabula
tesseract
pytesseract
openpyxl